+++ blocklets/ai-studio/src/components/file-editor/use-variables-editor-options.tsx
@@ -86,10 +86,11 @@ export default function useVariablesEditorOptions(assistant?: AssistantYjs) {
       doc.transact(() => {
         assistant.parameters ??= {};
 
-        if (!parameter || !variables.includes(parameter)) {
+        const key = (parameter || '').split('.')[0] || '';
+        if (!parameter || !variables.includes(key)) {
           assistant.parameters[id] = {
             index: Math.max(-1, ...Object.values(assistant.parameters).map((i) => i.index)) + 1,
-            data: { id, key: parameter, from },
+            data: { id, key, from },
           };
 
           setHighlightedId(id);
@@ -116,5 +117,19 @@ export default function useVariablesEditorOptions(assistant?: AssistantYjs) {
     [assistant]
   );
 
-  return { from, options, variables, addParameter, deleteParameter, highlightedId };
+  const removeParameter = useCallback(
+    (key: string) => {
+      if (!assistant) return;
+
+      const doc = (getYjsValue(assistant) as Map<any>).doc!;
+      doc.transact(() => {
+        if (!assistant.parameters) return;
+        for (const id of Object.keys(assistant.parameters)) {
+          if (assistant.parameters[id]?.data.key === key) delete assistant.parameters[id];
+        }
+      });
+    },
+    [assistant]
+  );
+  return { from, options, variables, addParameter, deleteParameter, removeParameter, highlightedId };
 }